+++ src/routes/graphql/types/context.ts
@@ -1,0 +1,7 @@
+import { PrismaClient } from '@prisma/client';
+import DataLoader from 'dataloader';
+
+export interface CtxType {
+  prisma: PrismaClient;
+  dataloaders: WeakMap<{ [k: string]: any }, DataLoader<string, unknown>>;
+}
